+++ plugins/numbers/tests/parse.test.js
@@ -12,17 +12,28 @@ test('parse numbers', function (t) {
     ['hundred and seventy two', 172],
     ['hundred and seventy two thousand', 172000],
     ['seven hundred and twenty', 720],
-    //
-    ['seven twenty', 720],
-    ['two twenty', 220],
-    ['two fifty', 250],
-    ['two fifty five', 255],
     ['three twenty', 320],
     ['hundred and three', 103],
+
+    //dropped-hundred
+    ['seven twenty', 720],
+    ['two twenty', 220],
+    ['eight fifty', 850],
+    ['two fifty', 250],
+    ['two fifty five', 255],
+
+    ['two and a half', 2.5],
+    ['fifty and a half', 50.5],
+    ['two fifty and three quarters', 250.75],
+    ['twenty two percent', 22],
+    ['two hundred and twenty percent', 220],
+    ['220 percent', 220],
+    ['$220', 220],
 
     // false-positives
     ['seven and twenty', 7],
     ['three three', 3],
+    ['four two zero', 4],
     // ['three twentieth', null],
     // ['third twenty', null],
     // ['seventy hundred', null],