+++ plugins/numbers/changelog.md
@@ -1,3 +1,8 @@
+### v1.2.0 [Feb 2021]
+- big improvements to `.fractions()` and decimal-parsing in the number-parser (thanks Jake!)
+- `.money()` parsing improvements -'35 cents' -> 0.35
+- add `.money().get()` method
+
 ### v1.1.0 [Dec 2020]
 - add `.money()` class methods:
   - `.money().currency()`